Refactor component imports and remove unused files
